% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_version}
\alias{fetch_version}
\alias{fetch_gshhg}
\title{Fetch a version and store in the root data path}
\usage{
fetch_version(
  url = "https://www.ngdc.noaa.gov/mgg/shorelines/data/gshhg/latest/gshhg-shp-2.3.7.zip",
  path = gshhg_path(),
  cleanup = TRUE
)

fetch_gshhg(
  url = "https://www.ngdc.noaa.gov/mgg/shorelines/data/gshhg/latest/gshhg-shp-2.3.7.zip",
  path = gshhg_path(),
  cleanup = TRUE
)
}
\arguments{
\item{url}{chr the URL to the version to download}

\item{path}{chr, the destination path}

\item{cleanup}{logical, if TRUE clean up the intermediary download files}
}
\value{
0 for success and non-zero otherwise
}
\description{
Fetch a version and store in the root data path
}
