% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gshhg-package.R
\docType{package}
\name{gshhg-package}
\alias{gshhg-package}
\alias{gshhg}
\title{gshhg: Managing GSHHG datasets with R}
\description{
Provides download, archiving and access to GSHHG related local datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/gshhg}
}

}
\keyword{internal}
