% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{gshhg_encodings}
\alias{gshhg_encodings}
\title{Retrieve tables of the filename encodings}
\usage{
gshhg_encodings()
}
\value{
named list of encoding tables
}
\description{
These are derived from the SHAPEFILE.TXT file
}
\details{
GSHHS_shp: Shapefiles (polygons) derived from shorelines. These are named 
`GSHHS_<resolution>_L<level>.*`
  
WDBII_shp: Political boundaries and rivers (lines). These are named
 `WDBII_border_<resolution>_L<level>.*` and `WDBII_river_<resolution>_L<level>.*`
}
