% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{gshhg_path}
\alias{gshhg_path}
\title{Retrieve the gshhg path}
\usage{
gshhg_path(..., root = root_path())
}
\arguments{
\item{...}{further arguments for \code{file.path()}}

\item{root}{the root path}
}
\value{
character path description
}
\description{
Retrieve the gshhg path
}
