% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gsshg_crop}
\alias{gsshg_crop}
\title{Crop a gshhg sf object}
\usage{
gsshg_crop(x, bb)
}
\arguments{
\item{x}{sf or sfc object}

\item{bb}{object from which an st_bbox can be extracted}
}
\value{
cropped geometry object (attributes of input are dropped)
}
\description{
Crop a gshhg sf object
}
