% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{help_shapefiles}
\alias{help_shapefiles}
\title{Read the "SHAPEFILE.TXT" for a given version}
\usage{
help_shapefiles(version = "2.3.7", path = gshhg_path())
}
\arguments{
\item{version}{chr, the data version (default 2.3.7)}

\item{path}{chr the gshhg data path}
}
\value{
chr vector of file data, one line per element
}
\description{
Read the "SHAPEFILE.TXT" for a given version
}
