% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_versions}
\alias{list_versions}
\title{List the available versions}
\usage{
list_versions(path = gshhg_path("versions"), ...)
}
\arguments{
\item{path}{the root data path}

\item{...}{other arguments for file.path}
}
\value{
character vector of version IDs
}
\description{
List the available versions
}
