% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{make_path}
\alias{make_path}
\title{Given a path - make it if it doesn't exist}
\usage{
make_path(path, recursive = TRUE, ...)
}
\arguments{
\item{path}{character, the path to check and/or create}

\item{recursive}{logical, create paths recursively?}

\item{...}{other arguments for \code{dir.create}}
}
\value{
the path
}
\description{
Given a path - make it if it doesn't exist
}
