% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gshhg}
\alias{read_gshhg}
\title{Read gshhg data by group, resolution and level.}
\usage{
read_gshhg(
  group = c("shoreline", "border", "river")[1],
  res = c("f", "h", "i", "l", "c")[1],
  level = 1,
  bb = NULL,
  version = "2.3.7",
  path = gshhg_path()
)
}
\arguments{
\item{group}{chr one of "shoreline" (default), "border" or "river"}

\item{res}{chr one of f, h, i, l or c (full, high, intermediate, low, crude). 
"f" is default.}

\item{level}{integer, for shoreline group 1-6, for border group 1-3 and for river group 1-11
Defaults to 1.}

\item{bb}{NULL or something from which a bounding box can be found for clipping,
if not NULL then attributes are dropped in favor of returning geometry only.}

\item{version}{chr, the data version (default 2.3.7)}

\item{path}{chr, the gshhg data path}
}
\value{
sf object
GSHHS_shp: Shapefiles (polygons) derived from shorelines. These are named 
`GSHHS_<resolution>_L<level>.*`
  
WDBII_shp: Political boundaries and rivers (lines). These are named
 `WDBII_border_<resolution>_L<level>.*` and `WDBII_river_<resolution>_L<level>.*`
}
\description{
See [`help_shapefiles`] for detailed explanations or 
[`gsshg_encodings`] for info on the encoding values.
}
